requirements.txt
PyYAML~=6.0
pandas~=2.0.0
scikit-learn~=1.2.0
optuna~=3.2.0
catboost~=1.2
numpy~=1.24.3
joblib~=1.2.0
uvicorn==0.22.0
fastapi~=0.96.0
scipy~=1.10.0
streamlit~=1.23.0
requests==2.31.0
seaborn~=0.12.0
matplotlib~=3.7.0
plotly~=5.15.0
fastparquet~=2023.4.0
